% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Brystbevarende_sh}
\alias{Brystbevarende_sh}
\title{Antallet med brystbevarende kirurgi og antallet potensielle med tumorstørrelse
0-30 mm pr. sykehus 2013-2015}
\format{To datarammer med 21 rader og 3 kolonner
\itemize{
  \item AntTilfeller: Radnavn er sykehusnavn, kolonnene er
   \itemize{
   \item 2013: Antall tilfeller i 2013
   \item 2014: Antall tilfeller i 2014
   \item 2015: Antall tilfeller i 2015
  }
  \item AntTotalt: Radnavn er sykehusnavn, kolonnene er
   \itemize{
   \item 2013: Antall tilfeller i 2013
   \item 2014: Antall tilfeller i 2014
   \item 2015: Antall tilfeller i 2015
  }
}}
\usage{
Brystbevarende_sh
}
\description{
En list som består av to datarammer, AntTilfeller og AntTotalt. Den første angir
antallet som har fått brystbevarende kirugi for alle aktuelle sykehus i årene
2013-2015, den andre angir antallet som var aktuelle for det.
}
\keyword{datasets}

