% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigAndelGrVar.R
\name{indikatorFigAndelGrVar}
\alias{indikatorFigAndelGrVar}
\title{Plot andeler/andeler i angitt format}
\usage{
indikatorFigAndelGrVar(AntTilfeller, N, outfile, tittel, width = 800,
  height = 700, sideTxt = "Boomr東e/opptaksomr東e", decreasing = F,
  terskel = 30, minstekrav = NA, maal = NA, skriftStr = 1.3,
  pktStr = 1.4, legPlass = "top", minstekravTxt = "Min.",
  maalTxt = "M胴", graaUt = NA)
}
\arguments{
\item{AntTilfeller}{En dataramme med antall i spesifisert form}

\item{N}{En dataramme med nevneren i andelsberegningen}

\item{outfile}{Angir filnavn og format p친 figuren som returneres,}

\item{graaUt}{En vektor med navn p친 enheter som skal ha gr친 s칮yler}
}
\value{
Et plot av andeler over tre 친r
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 친r,
der radnavn angir grupperingsvariabel og kolonnenavn 친r. Funksjonen
returnerer et s칮yleplot hvor s칮ylene representerer sist 친r, fyllt sirkel er 친ret
f칮r og 친pen sirkel to 친r f칮r
}

