% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigAndelGrVar_1aar.R
\name{indikatorFigAndelGrVar_1aar}
\alias{indikatorFigAndelGrVar_1aar}
\title{Plot alder og kj칮nnsjusterte andeler/andeler i angitt format}
\usage{
indikatorFigAndelGrVar_1aar(Antall, outfile, tittel, width = 800,
  height = 700, decreasing = F, terskel = 30, minstekrav = NA,
  minstekravTxt = "Min=", maal = NA, maalTxt = "M胴=", til100 = FALSE,
  skriftStr = 1.3, sideTxt = "Sykehus")
}
\arguments{
\item{outfile}{Angir filnavn og format p친 figuren som returneres,}

\item{andeler}{En dataramme med andeler/andeler i spesifisert form}

\item{N}{En vektor/matrise med N for ratene}
}
\value{
Et plot av andeler over tre 친r
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 친r,
der radnavn angir grupperingsvariabel og kolonnenavn 친r. Funksjonen
returnerer et s칮yleplot hvor s칮ylene representerer sist 친r, fyllt sirkel er 친ret
f칮r og 친pen sirkel to 친r f칮r
}

