% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigAndelGrVar_alderKjJust.R
\name{indikatorFigAndelGrVar_aldKjJust}
\alias{indikatorFigAndelGrVar_aldKjJust}
\title{Plot alder og kj칮nnsjusterte andeler/andeler i angitt format}
\usage{
indikatorFigAndelGrVar_aldKjJust(Antall, outfile, tittel, width = 800,
  height = 700, decreasing = F, terskel = 30, minstekrav = NA,
  minstekravTxt = "Min", sideTxt = "Boomr東e/opptaksomr東e", maal = NA,
  maalTxt = "M胴", til100 = FALSE, legPlass = "top", marker = NA)
}
\arguments{
\item{outfile}{Angir filnavn og format p친 figuren som returneres,}

\item{marker}{En vektor med navn p친 enheter som skal markeres med stjerne}

\item{AntTilfeller}{En dataramme med antall i spesifisert form}

\item{N}{En dataramme med nevneren i andelsberegningen}
}
\value{
Et plot av andeler over tre 친r
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 친r,
der radnavn angir grupperingsvariabel og kolonnenavn 친r. Funksjonen
returnerer et s칮yleplot hvor s칮ylene representerer sist 친r, fyllt sirkel er 친ret
f칮r og 친pen sirkel to 친r f칮r
}

