% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigAndelGrVar_alderKjJust_1aar.R
\name{indikatorFigAndelGrVar_aldKjJust_1aar}
\alias{indikatorFigAndelGrVar_aldKjJust_1aar}
\title{Plot alder og kjønnsjusterte andeler/andeler i angitt format}
\usage{
indikatorFigAndelGrVar_aldKjJust_1aar(Antall, outfile, tittel, width = 800,
  height = 700, decreasing = F, terskel = 30, minstekrav = NA,
  maal = NA, til100 = FALSE)
}
\arguments{
\item{outfile}{Angir filnavn og format på figuren som returneres,}

\item{andeler}{En dataramme med andeler/andeler i spesifisert form}

\item{N}{En vektor/matrise med N for ratene}
}
\value{
Et plot av andeler over tre år
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 år,
der radnavn angir grupperingsvariabel og kolonnenavn år. Funksjonen
returnerer et søyleplot hvor søylene representerer sist år, fyllt sirkel er året
før og åpen sirkel to år før
}

