% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigAndelGrVar_justert.R
\name{indikatorFigAndelGrVar_justert}
\alias{indikatorFigAndelGrVar_justert}
\title{Plot alder og kj칮nnsjusterte andeler/andeler i angitt format}
\usage{
indikatorFigAndelGrVar_justert(AntTilfeller, N, outfile, tittel, width = 800,
  height = 700, decreasing = F, terskel = 30, minstekrav = NA,
  sideTxt = "Boomr東e/opptaksomr東e", maal = NA, tertiler = c(-1, 51,
  66, 140), justeringLand = T)
}
\arguments{
\item{N}{En vektor/matrise med N for ratene}

\item{outfile}{Angir filnavn og format p친 figuren som returneres,}

\item{andeler}{En dataramme med andeler/andeler i spesifisert form}
}
\value{
Et plot av andeler over tre 친r
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 친r,
der radnavn angir grupperingsvariabel og kolonnenavn 친r. Funksjonen
returnerer et s칮yleplot hvor s칮ylene representerer sist 친r, fyllt sirkel er 친ret
f칮r og 친pen sirkel to 친r f칮r
}

