% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigAndelGrVar_preberegnet.R
\name{indikatorFigAndelGrVar_preberegnet}
\alias{indikatorFigAndelGrVar_preberegnet}
\title{Plot andeler/andeler i angitt format}
\usage{
indikatorFigAndelGrVar_preberegnet(Andeler, outfile, tittel, width = 800,
  height = 700, sideTxt = "Boomr東e/opptaksomr東e", decreasing = F,
  terskel = 30, minstekrav = NA, maal = NA, xtekst = "Andel \%",
  til100 = T, skriftStr = 1.3, pktStr = 1.5)
}
\arguments{
\item{outfile}{Angir filnavn og format p친 figuren som returneres,}

\item{andeler}{En dataramme med andeler/andeler i spesifisert form}

\item{N}{En vektor/matrise med N for ratene}
}
\value{
Et plot av andeler over tre 친r
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 친r,
der radnavn angir grupperingsvariabel og kolonnenavn 친r. Funksjonen
returnerer et s칮yleplot hvor s칮ylene representerer sist 친r, fyllt sirkel er 친ret
f칮r og 친pen sirkel to 친r f칮r
}

