% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigAntallGrVar.R
\name{indikatorFigAntallGrVar}
\alias{indikatorFigAntallGrVar}
\title{Plot antall i angitt format}
\usage{
indikatorFigAntallGrVar(Antall, outfile, tittel, width = 800, height = 700,
  sideTxt = "Boomr東e/opptaksomr東e", decreasing = F, xtekst = "Antall",
  skriftStr = 1.3, pktStr = 1)
}
\arguments{
\item{Antall}{En dataramme med antall i spesifisert form}

\item{outfile}{Angir filnavn og format p친 figuren som returneres,}
}
\value{
Et plot av antall over tre 친r
}
\description{
Denne funksjonen tar som input en dataramme med antall.Funksjonen
returnerer et s칮yleplot hvor s칮ylene representerer sist 친r, fyllt sirkel er 친ret
f칮r og 친pen sirkel to 친r f칮r
}

