% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indikatorFigRaterGrVar.R
\name{indikatorFigRaterGrVar}
\alias{indikatorFigRaterGrVar}
\title{Plot andeler/andeler i angitt format}
\usage{
indikatorFigRaterGrVar(Andeler, outfile, tittel, width = 600, height = 600,
  decreasing = F, terskel = 30, minstekrav = NA, maal = NA,
  xtekst = "Andel \%", til100 = F, skriftStr = 1.3, pktStr = 1.5)
}
\arguments{
\item{outfile}{Angir filnavn og format på figuren som returneres,}

\item{andeler}{En dataramme med andeler/andeler i spesifisert form}

\item{N}{En vektor/matrise med N for ratene}
}
\value{
Et plot av andeler over tre år
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 år,
der radnavn angir grupperingsvariabel og kolonnenavn år. Funksjonen
returnerer et søyleplot hvor søylene representerer sist år, fyllt sirkel er året
før og åpen sirkel to år før
}

