% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprosesser.R
\name{indikatorLastdataOgPreprosesser}
\alias{indikatorLastdataOgPreprosesser}
\title{Les og preprosesser data for fremstilling i figur}
\usage{
indikatorLastdataOgPreprosesser()
}
\value{
En rekke datasett organisert som forventet av funksjonene i IndikatoR
}
\description{
Denne funksjonenen laster dataen fra forskjellige kilder og
omformer den til formen forventet av funksjonene i IndikatoR
}

